Внес правки по задаче
