+++ src/HomeWork2/FindLongest.php
@@ -2,15 +2,17 @@
 
 class FindLongest
 {
-    public function findLongester(string $string)
+    public function findLongester($string)
     {
         $arr = explode(' ',$string);
-        foreach ($arr as $value=>$value) {
-        }
+        //var_dump($arr);
+        usort($arr,'mysort');
+    }
 
+    function mysort($a,$b){
+        return $a['0'] <=> $b['0'];
     }
 }
 
 $result = new FindLongest();
-$result->findLongester('def abc ghkeee');
-
+$result->findLongester('wefeirogh erwiogjweiorg ewrijgweoirgjrif');
