+++ tests/Homework1/GetSecondMaxTest.php
